♻️ Rename space folder
