+++ apps/web/src/app/[locale]/(space)/app-card.tsx
