+++ apps/web/src/app/[locale]/(space)/components/sidebar/app-sidebar-provider.tsx
