+++ apps/web/src/app/[locale]/(space)/components/sidebar/nav-item.tsx
