+++ apps/web/src/app/[locale]/(space)/components/sidebar/nav-user.tsx
