+++ apps/web/src/app/[locale]/(space)/components/top-bar.tsx
