+++ apps/web/src/app/[locale]/(space)/events/past-events.tsx
@@ -1,7 +1,6 @@
 "use client";
 import { CalendarPlusIcon } from "lucide-react";
 
-import { EventList } from "@/app/[locale]/(admin)/events/event-list";
 import {
   EmptyState,
   EmptyStateDescription,
@@ -12,8 +11,12 @@ import { Spinner } from "@/components/spinner";
 import { Trans } from "@/components/trans";
 import { trpc } from "@/trpc/client";
 
-export function UpcomingEvents() {
-  const { data } = trpc.scheduledEvents.list.useQuery({ period: "upcoming" });
+import { EventList } from "./event-list";
+
+export function PastEvents() {
+  const { data } = trpc.scheduledEvents.list.useQuery({
+    period: "past",
+  });
 
   if (!data) {
     return <Spinner />;
@@ -27,13 +30,13 @@ export function UpcomingEvents() {
         </EmptyStateIcon>
         <EmptyStateTitle>
           <Trans
-            i18nKey="upcomingEventsEmptyStateTitle"
-            defaults="No Upcoming Events"
+            i18nKey="pastEventsEmptyStateTitle"
+            defaults="No Past Events"
           />
         </EmptyStateTitle>
         <EmptyStateDescription>
           <Trans
-            i18nKey="upcomingEventsEmptyStateDescription"
+            i18nKey="pastEventsEmptyStateDescription"
             defaults="When you schedule events, they will appear here."
           />
         </EmptyStateDescription>
