+++ apps/web/src/app/[locale]/(space)/loading.tsx
