+++ apps/web/src/app/[locale]/(space)/mobile-navigation.tsx
