+++ apps/web/src/app/[locale]/(space)/page.tsx
