+++ apps/web/src/app/[locale]/(space)/polls/polls-tabbed-view.tsx
