+++ apps/web/src/app/[locale]/(space)/recently-updated-polls-grid.tsx
