+++ apps/web/src/app/[locale]/(space)/settings/components/settings-layout.tsx
