+++ apps/web/src/app/[locale]/(space)/settings/components/sign-out-button.tsx
