+++ apps/web/src/app/[locale]/(space)/settings/layout.tsx
