+++ apps/web/src/app/[locale]/(space)/settings/loading.tsx
