+++ apps/web/src/app/[locale]/(space)/settings/preferences/loading.tsx
