+++ apps/web/src/app/[locale]/(space)/settings/profile/delete-account-dialog.tsx
