+++ apps/web/src/app/[locale]/(space)/settings/profile/profile-page.tsx
@@ -3,16 +3,16 @@ import { Button } from "@rallly/ui/button";
 import { DialogTrigger } from "@rallly/ui/dialog";
 import { TrashIcon } from "lucide-react";
 
-import {
-  SettingsContent,
-  SettingsSection,
-} from "@/app/[locale]/(admin)/settings/components/settings-layout";
-import { DeleteAccountDialog } from "@/app/[locale]/(admin)/settings/profile/delete-account-dialog";
-import { ProfileSettings } from "@/app/[locale]/(admin)/settings/profile/profile-settings";
 import { Trans } from "@/components/trans";
 import { useUser } from "@/components/user-provider";
 
+import {
+  SettingsContent,
+  SettingsSection,
+} from "../components/settings-layout";
+import { DeleteAccountDialog } from "./delete-account-dialog";
 import { ProfileEmailAddress } from "./profile-email-address";
+import { ProfileSettings } from "./profile-settings";
 
 export const ProfilePage = () => {
   const { user } = useUser();
