+++ apps/web/src/app/[locale]/(space)/settings/profile/profile-picture.tsx
